Implement Todo BAD_REQUEST Response Functionality
Implement the functionality to respond to bad request made to the todo
endpoint.

Add Package Support For uuid

Create Test For Status Code BAD_REQUEST

  1. Setup jest to reset mocks after each test.

     Status Quo: The same mocks are shared by all test cases.

     Problem: This could potentially result in one test interfaring with
     the resulst of another test.

     Solution: Clear all mocks after running each individual test.

  2. Create test case to verify that the "TodoController.get_single"
     function responds with status code "BAD REQUEST" when an invalid
     id is given.

Update Test Case Group Names

Status Quo: The name given to the describe function in the test for
TodoController.get_single is "get".

Problem: "describe block" names should be the name of the function under
test.

Solution: Change the name to "get_single"

Add Package Support For uuid types

Create Validation Functionality For Id ValueObject

Status Quo: Id value obejct does not have any validation implemented.

Problem: We want to use uuid for Ids to ensure uniqueness. This needs to
be enforced by some kind of validation.

Solution: Create a function that validates id value object DTOs.
Ensuring that the given dto is a valide uuid and the appropriate type.

Implement BAD_REQUEST Reason Phrase Functionality

Status Quo: We do not have any way of verifying that the todo api
consumer is sending a valid id when requesting (GET) a todo item.

Problem: We need to be able to let the api consumer know whether an
error is due to their input.

Solution: Implement the functionality to verify the id portion of the
route sent by the todo api consumer.