Lift INTERNAL_SERVER_ERROR Mock Function
Status Quo: The mocked version of Repository.Todo.find in the tests for
INTERNAL_SERVER_ERROR, are in the do_test function for each test.

Problem: This could create multiple mocks for the same function doing
the same thing.

Solution: Move the mocked function to the scope of the "it()" functions
callback scope.