Setup Todo Entity Api Response Code NOT FOUND
Status Quo: If the todo item requested by the api consumer does not
exist, we simply don't return anything.

Problem: We need to let the api consumer know whenever something has
happend. In this case, we need to let them know that the item they are
looking for was not found so that they may take appropriate actions.

Solution: Update TodoController.get_single to return a "NOT FOUND"
status code whenever the requested todo item cannot be found.