Setup Todo Entity Api Response Phrase NOT FOUND
Status Quo: the TodoController.get_single handler is set up to respond
with the appropriate status code if the requested todo item cannot be
found, but no reason phrase is being sent to the api consumer.

Problem: Best practice is to respond with both a status code and a
reason phrase.

Solution: update TodoController.get_single to respond with the reason
phrase corresponding to the "NOT FOUND" status code.