Validate Todo Repository DTO
Status Quo: start function in the main api server file is no validating
the todo repository data transfer object before passing it to the todo
router initialization function.

Problem: The todo router initialization function assumes that the given
todo repository has be validated. If repository is not validated, this
could result in errors down the line.

Solution: update the start function to valiate the repository before
passing it to the initialization function (i.e.
TodoRouter.instanciate)